+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added the `commerce/gateways/webhook-url` command.
 - Fixed a bug where the delete button would be shown for users that do not have permission to delete on the Product edit page. ([#3285](https://github.com/craftcms/commerce/issues/3285))
 - Fixed a bug where deleted shipping categories were still available for selection. ([#3272](https://github.com/craftcms/commerce/issues/3272))
+- Fixed a bug where the customer condition rule wasn’t loading correctly. ([#3291](https://github.com/craftcms/commerce/issues/3291))
 
 ## 4.3.0 - 2023-09-13